renamed other files to correct naming
