+++ sorting/selection_sort/ascSelectionSortWithMin.c
