+++ sorting/selection_sort/descSelectionSortUsingMax.c
@@ -1,4 +1,4 @@
-//INSERTION SORT
+//SELECTION SORT
 //USING MAX ELEMENT
 //DESCENDING
 #include <stdio.h>
