+++ sorting/shell_sort/ascShellSort.java
@@ -1,5 +1,5 @@
 import java.util.*;
-public class ShellSort
+public class ascShellSort
 {
     public static void main(String[] args)
     {
