+++ strings/anagram/anagram.py
