+++ strings/pallindrome/pallindrome.c
