+++ strings/pallindrome/pallindrome.cpp
