+++ strings/pattern_searching/z/z.cpp
